refactor: listing images
Refactor of the database helper that had some image related functions. Passed them to the image helper where they should be.
